stage_2_pattern_extract/requirements.txt
# Stage 2: Pattern Extract
# Virtual environment: nemo_data_designer
#
# Setup (Python 3.12 required):
#   python3.12 -m venv .venv
#   source .venv/bin/activate
#   pip install -r requirements.txt
#
# Note: data-designer package may require NVIDIA NIM credentials.
# The script connects to a local vLLM/NIM server at http://localhost:8000/v1

annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.0
colorama==0.4.6
data-designer==0.5.7
distro==1.9.0
filelock==3.29.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.5.0
httpx-sse==0.4.3
huggingface_hub==1.11.0
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
json_repair==0.59.4
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
networkx==3.6.1
numpy==2.4.4
openai==2.32.0
packaging==25.0
pandas==2.3.3
platformdirs==4.9.6
prompt_toolkit==3.0.52
pyarrow==19.0.1
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==14.3.4
rpds-py==0.30.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
